Checkin of major mods for core allocation logic. The past month work for Chris Holmes   and Moe Jette. There are still a couple of end-cases that need to be fixed. We also   need to push the core identification out to the slurmd for task binding. sched/gang   needs revision to support this too.

+++ trunk/src/common/env.h
@@ -260,8 +260,8 @@ char **env_array_user_default(const char *username, int timeout, int mode);
 /*
  * The cpus-per-node representation in SLURM (and perhaps tasks-per-node
  * in the future) is stored in a compressed format comprised of two
- * equal-length arrays of uint32_t, and an integer holding the array length.
- * In one array an element represents a count (number of cpus, number of tasks,
+ * equal-length arrays, and an integer holding the array length. In one 
+ * array an element represents a count (number of cpus, number of tasks,
  * etc.), and the corresponding element in the other array contains the
  * number of times the count is repeated sequentially in the uncompressed
  * something-per-node array.
@@ -270,7 +270,7 @@ char **env_array_user_default(const char *username, int timeout, int mode);
  * array.  Free with xfree().
  */
 char *uint32_compressed_to_str(uint32_t array_len,
-			       const uint32_t *array,
+			       const uint16_t *array,
 			       const uint32_t *array_reps);
 
 #endif
